+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [1.0.0] - [2018/05/25]    
     * Basic Usage
     
-
+## [1.0.1] - [2018/05/25]    
+    * Fix index change bug when switchStyle=DropdownMenuShowHideSwitchStyle.directHideAnimationShow
+    * Add index change swith animation style : DropdownMenuShowHideSwitchStyle.directHideDirectShow